{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/urls.py
from django.conf.urls import url, include
from django.contrib import admin

from rest_framework import routers


router = routers.DefaultRouter()

urlpatterns = [
    url(r'^admin/', admin.site.urls),

    url(r'', include(router.urls)),
]
