{{cookiecutter.repo_name}}/requirements/requirements.txt
# Minimal requirements to run the {{cookiecutter.repo_name}} application in production
# Conventions:
#  - Always pin requirements exactly (using ==). This prevents unexpected upgrades breaking the code
#  - Always include dependencies of packages. We install with --no-deps to prevent version clashes between dependencies
#  - Put an empty line and comment line above each requirement, with the dependencies directly underneath
Django==1.11.6

# REST API
djangorestframework==3.7.3

# Work with timezones
pytz==2017.3

# Compatibility between Python 2 and 3
six==1.11.0

# Shell plus
django-extensions==1.9.7

# Use environment variables for configuration
django-environ==0.4.4

# Mysql database support
mysqlclient==1.3.12

# Translation of models in the admin
django-parler==1.7

# Image processing: database image fields
Pillow==4.3.0
olefile==0.44
